Group Creation
Groups which also included 1-1 chats can now be made. There is still a token error which needs to be fixed and seems to be a backend error. Will be solved in the next commit.

Backend code was also updated to have UserController which handles the user related apis.

GroupCreateModal introduced for creating new groups and GroupLists as well which still needs to be tested.