Messaging Introduced
Messaging can now take place without any errors. There is the issue of retrieving older messages and displaying the current ones according to the timestamps which needs to be done by the next commit. Plus the MessageInput was made better with emojis and file input (this cannot be checked rn tho).