Fixed the Message Duplication Error
Optimistic Messages which were being declared in the webSocket are now declared and dealt with in the ChatContainer. So the messages were duplicated due to the optimistic messages (in the local messages) plus them also being in the realTime ones as well. So we now clear them up once we change states.